+++ test/unit/internal/test_trust.py
@@ -21,13 +21,39 @@ from cryptography.hazmat.primitives.serialization import Encoding, PublicFormat
 from cryptography.x509 import load_pem_x509_certificate
 from sigstore_protobuf_specs.dev.sigstore.common.v1 import TimeRange
 
-from sigstore._internal.trustroot import (
+from sigstore._internal.trust import (
+    ClientTrustConfig,
     KeyringPurpose,
     TrustedRoot,
     _is_timerange_valid,
 )
 from sigstore._utils import load_pem_public_key
-from sigstore.errors import RootError
+from sigstore.errors import Error, RootError
+
+
+class TestTrustedRoot:
+    def test_good(self, asset):
+        path = asset("trusted_root/trustedroot.v1.json")
+        root = TrustedRoot.from_file(path)
+
+        assert (
+            root._inner.media_type == TrustedRoot.TrustedRootType.TRUSTED_ROOT_0_1.value
+        )
+        assert len(root._inner.tlogs) == 1
+        assert len(root._inner.certificate_authorities) == 2
+        assert len(root._inner.ctlogs) == 2
+        assert len(root._inner.timestamp_authorities) == 1
+
+    def test_bad_media_type(self, asset):
+        path = asset("trusted_root/trustedroot.badtype.json")
+
+        with pytest.raises(
+            Error, match="unsupported trusted root format: bad-media-type"
+        ):
+            TrustedRoot.from_file(path)
+
+
+# TODO(ww): Move these into appropriate class-scoped tests.
 
 
 def test_trust_root_tuf_caches_and_requests(mock_staging_tuf, tuf_dirs):
@@ -37,7 +63,7 @@ def test_trust_root_tuf_caches_and_requests(mock_staging_tuf, tuf_dirs):
     # keep track of requests the TrustUpdater invoked by TrustedRoot makes
     reqs, fail_reqs = mock_staging_tuf
 
-    trust_root = TrustedRoot.staging(purpose=KeyringPurpose.VERIFY)
+    trust_root = TrustedRoot.staging()
     # metadata was "downloaded" from staging
     expected = ["root.json", "snapshot.json", "targets.json", "timestamp.json"]
     assert sorted(os.listdir(data_dir)) == expected
@@ -53,15 +79,15 @@ def test_trust_root_tuf_caches_and_requests(mock_staging_tuf, tuf_dirs):
     assert reqs == expected_requests
     assert fail_reqs == expected_fail_reqs
 
-    trust_root.ct_keyring()
-    trust_root.rekor_keyring()
+    trust_root.ct_keyring(KeyringPurpose.VERIFY)
+    trust_root.rekor_keyring(KeyringPurpose.VERIFY)
 
     # no new requests
     assert reqs == expected_requests
     assert fail_reqs == expected_fail_reqs
 
     # New trust root (and TrustUpdater instance), same cache dirs
-    trust_root = TrustedRoot.staging(purpose=KeyringPurpose.VERIFY)
+    trust_root = TrustedRoot.staging()
 
     # Expect new timestamp and root requests
     expected_requests["timestamp.json"] += 1
@@ -69,8 +95,8 @@ def test_trust_root_tuf_caches_and_requests(mock_staging_tuf, tuf_dirs):
     assert reqs == expected_requests
     assert fail_reqs == expected_fail_reqs
 
-    trust_root.ct_keyring()
-    trust_root.rekor_keyring()
+    trust_root.ct_keyring(purpose=KeyringPurpose.VERIFY)
+    trust_root.rekor_keyring(purpose=KeyringPurpose.VERIFY)
     # Expect no requests
     assert reqs == expected_requests
     assert fail_reqs == expected_fail_reqs
@@ -83,7 +109,7 @@ def test_trust_root_tuf_offline(mock_staging_tuf, tuf_dirs):
     # keep track of requests the TrustUpdater invoked by TrustedRoot makes
     reqs, fail_reqs = mock_staging_tuf
 
-    trust_root = TrustedRoot.staging(offline=True, purpose=KeyringPurpose.VERIFY)
+    trust_root = TrustedRoot.staging(offline=True)
 
     # Only the embedded root is in local TUF metadata, nothing is downloaded
     expected = ["root.json"]
@@ -91,8 +117,8 @@ def test_trust_root_tuf_offline(mock_staging_tuf, tuf_dirs):
     assert reqs == {}
     assert fail_reqs == {}
 
-    trust_root.ct_keyring()
-    trust_root.rekor_keyring()
+    trust_root.ct_keyring(purpose=KeyringPurpose.VERIFY)
+    trust_root.rekor_keyring(purpose=KeyringPurpose.VERIFY)
 
     # Still no requests
     assert reqs == {}
@@ -153,35 +179,71 @@ def test_trust_root_bundled_get(monkeypatch, mock_staging_tuf, tuf_asset):
     ]
 
     # Assert that trust root from TUF contains the expected keys/certs
-    trust_root = TrustedRoot.staging(purpose=KeyringPurpose.VERIFY)
+    trust_root = TrustedRoot.staging()
     assert ctfe_keys[0] in get_public_bytes(
-        [k.key for k in trust_root.ct_keyring()._keyring.values()]
+        [
+            k.key
+            for k in trust_root.ct_keyring(
+                purpose=KeyringPurpose.VERIFY
+            )._keyring.values()
+        ]
     )
     assert (
-        get_public_bytes([k.key for k in trust_root.rekor_keyring()._keyring.values()])
+        get_public_bytes(
+            [
+                k.key
+                for k in trust_root.rekor_keyring(
+                    purpose=KeyringPurpose.VERIFY
+                )._keyring.values()
+            ]
+        )
         == rekor_keys
     )
     assert trust_root.get_fulcio_certs() == fulcio_certs
 
     # Assert that trust root from offline TUF contains the expected keys/certs
-    trust_root = TrustedRoot.staging(offline=True, purpose=KeyringPurpose.VERIFY)
+    trust_root = TrustedRoot.staging(offline=True)
     assert ctfe_keys[0] in get_public_bytes(
-        [k.key for k in trust_root.ct_keyring()._keyring.values()]
+        [
+            k.key
+            for k in trust_root.ct_keyring(
+                purpose=KeyringPurpose.VERIFY
+            )._keyring.values()
+        ]
     )
     assert (
-        get_public_bytes([k.key for k in trust_root.rekor_keyring()._keyring.values()])
+        get_public_bytes(
+            [
+                k.key
+                for k in trust_root.rekor_keyring(
+                    purpose=KeyringPurpose.VERIFY
+                )._keyring.values()
+            ]
+        )
         == rekor_keys
     )
     assert trust_root.get_fulcio_certs() == fulcio_certs
 
     # Assert that trust root from file contains the expected keys/certs
     path = tuf_asset.target_path("trusted_root.json")
-    trust_root = TrustedRoot.from_file(path, purpose=KeyringPurpose.VERIFY)
+    trust_root = TrustedRoot.from_file(path)
     assert ctfe_keys[0] in get_public_bytes(
-        [k.key for k in trust_root.ct_keyring()._keyring.values()]
+        [
+            k.key
+            for k in trust_root.ct_keyring(
+                purpose=KeyringPurpose.VERIFY
+            )._keyring.values()
+        ]
     )
     assert (
-        get_public_bytes([k.key for k in trust_root.rekor_keyring()._keyring.values()])
+        get_public_bytes(
+            [
+                k.key
+                for k in trust_root.rekor_keyring(
+                    purpose=KeyringPurpose.VERIFY
+                )._keyring.values()
+            ]
+        )
         == rekor_keys
     )
     assert trust_root.get_fulcio_certs() == fulcio_certs
@@ -194,15 +256,35 @@ def test_trust_root_tuf_instance_error():
 
 def test_trust_root_tuf_ctfe_keys_error(monkeypatch):
     trust_root = TrustedRoot.staging(offline=True)
-    monkeypatch.setattr(trust_root, "ctlogs", [])
+    monkeypatch.setattr(trust_root._inner, "ctlogs", [])
     with pytest.raises(Exception, match="CTFE keys not found in trusted root"):
-        trust_root.ct_keyring()
+        trust_root.ct_keyring(purpose=KeyringPurpose.VERIFY)
 
 
 def test_trust_root_fulcio_certs_error(tuf_asset, monkeypatch):
     trust_root = TrustedRoot.staging(offline=True)
-    monkeypatch.setattr(trust_root, "certificate_authorities", [])
+    monkeypatch.setattr(trust_root._inner, "certificate_authorities", [])
     with pytest.raises(
         Exception, match="Fulcio certificates not found in trusted root"
     ):
         trust_root.get_fulcio_certs()
+
+
+class TestClientTrustConfig:
+    def test_good(self, asset):
+        path = asset("trust_config/config.v1.json")
+        config = ClientTrustConfig.from_json(path.read_text())
+
+        assert config._inner.signing_config.ca_url == "https://fakeca.example.com"
+        assert config._inner.signing_config.oidc_url == "https://fakeoidc.example.com"
+        assert config._inner.signing_config.tlog_urls == ["https://fakelog.example.com"]
+        assert config._inner.signing_config.tsa_urls == ["https://faketsa.example.com"]
+        assert isinstance(config.trusted_root, TrustedRoot)
+
+    def test_bad_media_type(self, asset):
+        path = asset("trust_config/config.badtype.json")
+
+        with pytest.raises(
+            Error, match="unsupported client trust config format: bad-media-type"
+        ):
+            ClientTrustConfig.from_json(path.read_text())